Replace nodes in second pass after loading variables.
Before there was a bit of a race condition with regards to async resolvers and loaders – if a resolver didn't resolve synchronously the walk would continue and variables would get replaced with `undefined` before their real values were actually loaded. Now there are two separate phases – one for the resolution and one for the replacement, ensuring all values are resolved before they're replaced.